requirement.txt
langchain 
chromadb 
langchain-openai 
pypdf 
gradio
groq
python-dotenv